Move fairing to fairing.rs module
